Change samplerate_t type to double

+++ src/mojo/core/audio/types.hpp
@@ -9,7 +9,7 @@ namespace mojo {
 
 typedef int64_t count_t;
 typedef uint32_t channel_count_t;
-typedef uint32_t samplerate_t;
+typedef double samplerate_t;
 
 } // namespace mojo
